server/sell-script-backend/requirements.txt
Django==1.9
django-bootstrap3==6.2.2
djangorestframework==3.3.2
gunicorn==19.4.4
django-statici18n==1.1.5
Pillow==3.1.1
psycopg2==2.6.1
html2text==2016.1.8
django-compressor==2.0
python-social-auth==0.2.14
drf-extensions==0.2.8
drf-nested-routers==0.11.1
